💈🐩 -> Confirmed that traditional english spelling was part of the cause of #32
