+++ package.json
@@ -6,9 +6,12 @@
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
+    "@thirdweb-dev/react": "^2.4.1",
+    "@thirdweb-dev/sdk": "^2.2.11",
     "antd": "^4.18.7",
     "assert": "^2.0.0",
     "crypto-browserify": "^3.12.0",
+    "ethers": "^5.6.6",
     "https-browserify": "^1.0.0",
     "moralis": "^1.3.2",
     "os-browserify": "^0.3.0",
