--- src/sampleProposals.js
@@ -1,29 +1,0 @@
-[
-    [
-      1,
-      <div>Should we start a Moralis hamburger chain?</div>,
-      <Tag color="green" text="Passed" />,
-    ],
-    [
-      2,
-      "Should we accept Elon Musks $44billion offer for our DAO?",
-      <Link to="/proposal" state={"hello"}>
-        <Tag color="red" text="Rejected" />
-      </Link>,
-    ],
-    [
-      3,
-      "Do you want a Web3 Slack tutorial?",
-      <Tag color="blue" text="Ongoing" />,
-    ],
-    [
-      4,
-      "Are you interested in Xbox/Console web3 tutorials?",
-      <Tag color="blue" text="Ongoing" />,
-    ],
-    [
-      5,
-      "Would you attend a Moralis Builder get together in Miami?",
-      <Tag color="blue" text="Ongoing" />,
-    ],
-]
